3 pratice
1.Mother's Milk
Farmer John has three milking buckets of capacity A, B, and C liters.
Each of the numbers A, B, and C is an integer from 1 through 20,
inclusive. Initially, buckets A and B are empty while bucket C is full
of milk. Sometimes, FJ pours milk from one bucket to another until the
second bucket is filled or the first bucket is empty. Once begun, a pour
must be completed, of course. Being thrifty, no milk may be tossed out.
Write a program to help FJ determine what amounts of milk he can leave
in bucket C when he begins with three buckets as above, pours milk among
the buckets for a while, and then notes that bucket A is empty.
PROGRAM NAME: milk3
INPUT FORMAT
A single line with the three integers A, B, and C.
SAMPLE INPUT (file milk3.in)
8 9 10
OUTPUT FORMAT
A single line with a sorted list of all the possible amounts of milk
that can be in bucket C when bucket A is empty.
SAMPLE OUTPUT (file milk3.out)
1 2 8 9 10
SAMPLE INPUT (file milk3.in)
2 5 10
SAMPLE OUTPUT (file milk3.out)
5 6 7 8 9 10

2. Number Triangles
Consider the number triangle shown below. Write a program that
calculates the highest sum of numbers that can be passed on a route that
starts at the top and ends somewhere on the base. Each step can go
either diagonally down to the left or diagonally down to the right.
7
3   8
8   1   0
2   7   4   4
4   5   2   6   5
In the sample above, the route from 7 to 3 to 8 to 7 to 5 produces the
highest sum: 30.
PROGRAM NAME: numtri
INPUT FORMAT
The first line contains R (1 <= R <= 1000), the number of rows. Each
subsequent line contains the integers for that particular row of the
triangle. All the supplied integers are non-negative and no larger than
100.
SAMPLE INPUT (file numtri.in)
5
7
3 8
8 1 0
2 7 4 4
4 5 2 6 5
OUTPUT FORMAT
A single line containing the largest sum using the traversal specified.
SAMPLE OUTPUT (file numtri.out)
30

3. Prime Palindromes
The number 151 is a prime palindrome because it is both a prime number
and a palindrome (it is the same number when read forward as backward).
Write a program that finds all prime palindromes in the range of two
supplied numbers a and b (5 <= a < b <= 100,000,000); both a and b are
considered to be within the range .
PROGRAM NAME: pprime
INPUT FORMAT
Line 1:
Two integers, a and b
SAMPLE INPUT (file pprime.in)
5 500
OUTPUT FORMAT
The list of palindromic primes in numerical order, one per line.
SAMPLE OUTPUT (file pprime.out)
5
7
11
101
131
151
181
191
313
353
373
383